Moving TZ database to app test resources
